+++ saber/__init__.py
@@ -1,13 +1,13 @@
 import saber.assign
 import saber.bs
-import saber.calibrate
+import saber.saber
 import saber.cluster
 import saber.gis
 import saber.io
 import saber.table
 
 __all__ = [
-    'io', 'table', 'cluster', 'assign', 'gis', 'calibrate', 'bs.py',
+    'io', 'table', 'cluster', 'assign', 'gis', 'saber.py', 'bs.py',
 ]
 
 __author__ = 'Riley C. Hales'
